requirements.txt
asgiref==3.7.2
celery==5.3.1
flower==2.0.0
fastapi==0.100.0
pydantic==2.0.3
uvicorn==0.23.0
redis==4.6.0
amqp==5.1.1
jsonpickle==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow ==10.0.0